move aggregation selection logic to metric reader
