+++ packages/solarwinds-apm/src/appoptics/exporters/metrics.ts
@@ -23,11 +23,9 @@ import {
   type InstrumentationScope,
 } from "@opentelemetry/core"
 import {
-  Aggregation,
   AggregationTemporality,
   DataPointType,
   type ExponentialHistogram as ExponentialHistogramDatapoint,
-  ExponentialHistogramAggregation,
   type Histogram as HistogramDatapoint,
   InstrumentType,
   type MetricDescriptor,
@@ -142,16 +140,6 @@ export class AppopticsMetricExporter {
     })
   }
 
-  selectAggregation(instrumentType: InstrumentType): Aggregation {
-    switch (instrumentType) {
-      case InstrumentType.HISTOGRAM: {
-        return new ExponentialHistogramAggregation(undefined, true)
-      }
-      default: {
-        return Aggregation.Default()
-      }
-    }
-  }
   selectAggregationTemporality(
     instrumentType: InstrumentType,
   ): AggregationTemporality {
