+++ packages/solarwinds-apm/src/exporters/metrics.ts
@@ -20,6 +20,8 @@ import {
   AggregationTemporality,
   ExponentialHistogramAggregation,
   InstrumentType,
+  PeriodicExportingMetricReader,
+  type PeriodicExportingMetricReaderOptions,
 } from "@opentelemetry/sdk-metrics"
 
 import { type Configuration } from "../shared/config.js"
@@ -34,6 +36,22 @@ export class MetricExporter extends OTLPMetricExporter {
       },
     })
   }
+}
+
+const DEFAULT_CARDINALITY_LIMIT = 200
+export interface MetricReaderOptions
+  extends PeriodicExportingMetricReaderOptions {
+  cardinalityLimit?: number
+}
+
+export class MetricReader extends PeriodicExportingMetricReader {
+  readonly #cardinalityLimit: number
+
+  constructor(options: MetricReaderOptions) {
+    super(options)
+    this.#cardinalityLimit =
+      options.cardinalityLimit ?? DEFAULT_CARDINALITY_LIMIT
+  }
 
   override selectAggregation(instrumentType: InstrumentType): Aggregation {
     switch (instrumentType) {
@@ -47,5 +65,12 @@ export class MetricExporter extends OTLPMetricExporter {
   }
   override selectAggregationTemporality(): AggregationTemporality {
     return AggregationTemporality.DELTA
+  }
+
+  override selectCardinalityLimit(instrumentType: InstrumentType): number {
+    return Math.min(
+      super.selectCardinalityLimit(instrumentType),
+      this.#cardinalityLimit,
+    )
   }
 }
